move NameConverter test logic to FunctionTest
