+++ Tests/Functional/Controller/ApiController80.php
@@ -18,6 +18,7 @@ use Nelmio\ApiDocBundle\Annotation\Security;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\Article;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\ArticleInterface;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\CompoundEntity;
+use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityThroughNameConverter;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithAlternateType80;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithNullableSchemaSet;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithObjectType;
@@ -466,6 +467,20 @@ class ApiController80
      * @Route("/serializename", methods={"GET"})
      */
     public function serializedNameAction()
+    {
+    }
+
+    /**
+     * @Route("/name_converter_context", methods={"GET"})
+     *
+     * @OA\Response(
+     *    response="200",
+     *    description="",
+     *
+     *    @Model(type=EntityThroughNameConverter::class, serializationContext={"secret_name_converter_value"=true})
+     * )
+     */
+    public function nameConverterContext()
     {
     }
 }
